hevc/gpu: enable 16-bit shader features, swap ref buffers, harden two sizes
Continuing the simpmix/bc250-encoding-decoding-fix fork review (backlog
section G, G7-G9). Three independent, individually-verified changes:

gpu_compute.c: deblock_filter/intra_wavefront/quantize/reconstruct all
declare SPIR-V Int16/StorageBuffer16BitAccess, but bc250_gpu_init()'s
vkCreateDevice() only ever requested timelineSemaphore - shaderInt16,
storageBuffer16BitAccess and uniformAndStorageBuffer16BitAccess were never
enabled. Using a capability without enabling the matching device feature
is undefined behaviour by the Vulkan spec: the fork that flagged this
reports a real SEGV inside libvulkan_radeon.so and a 5.3 dB "almost black"
HEVC corruption from the same bug on the same GPU family. Fixed by
querying vkGetPhysicalDeviceFeatures2/VkPhysicalDeviceVulkan11Features for
what the device actually offers and chaining that into
VkDeviceCreateInfo.pNext, falling back to a stderr warning (not a hard
failure) if a future device lacks it. Purely additive, so no regression is
possible; on-board confirmation still pending since this only touches
Vulkan device creation and shader dispatch, which host-drift can't reach.

encoder_h265.c: encode_core() used to memcpy recon_y/cb/cr into
prev_recon_y/cb/cr every frame (~3MB/frame at 1080p). prev_recon_* is only
ever read and recon_* is only ever written in full each frame, so the
buffers can just change places - both are separately malloc()'d at the
same size and freed together, confirmed before making the change. First
attempt placed the swap before the BC250_HEVC_DEBUG_RECON debug dump,
which names the buffer by role (encoder->recon_y) rather than by frame -
after an early swap that name pointed at the previous frame's data.
hevc_host_drift.sh caught it immediately (53/53 -> real byte differences on
every case). Fixed by moving the swap after the dump. Re-verified 53/53.

encoder_h264.c + gpu_compute.c: two buffer-size computations
(output_buf_size = width * height * 4 + 65536, entropy_size = width *
height * 2) multiplied in 32-bit before widening to size_t/VkDeviceSize -
harmless at every size this driver accepts today (max 4096x4096) but wrong
in principle. Cast before multiplying, matching how every other buffer
size in this codebase is computed. A third instance from the same CodeQL
pass (a NAL-header buffer over-read in a standalone hevcps.c tool) doesn't
apply - that tool isn't part of this tree.

Verified off-board: hevc_host_drift.sh 53/53 byte-exact, ctest 6/6 relevant
suites (VaApiDriverTest pre-existingly fails off-board).

docs/backlog.md section G also records G5/G6 (already fixed or not
applicable - HEVC rate-control gating, merge-candidate z-scan rewrite, POC
LSB width, the EndPicture DRIVER_LOCK-drop idea flagged for dedicated
future review) and a reviewed-not-applicable list (VA-API RC-mode-default
churn, a packed-headers change that contradicts this project's own
documented reasoning, a GPU shader SAD-reduction optimization needing
board time to validate safely, and a CPU quantization optimization already
implemented here more rigorously).

+++ approach1-compute-encoder/src/encoder_h264.c
@@ -1731,7 +1731,7 @@ h264_encoder_t *h264_encoder_create(bc250_gpu_context_t *gpu_ctx,
      * below). At 1440p this is 14.8 MB vs 7.4 MB - an irrelevant amount of
      * host memory for one encoder instance - and it took the pathological
      * case from 12-of-12 frames refused to 0. */
-    encoder->output_buf_size = width * height * 4 + 65536;
+    encoder->output_buf_size = (size_t)width * height * 4 + 65536;
     encoder->output_buf = malloc(encoder->output_buf_size);
     if (!encoder->output_buf) {
         free(encoder);
